Split out policy JSON loading from view
