+++ app/js/actions/PolicyActions.js
@@ -1,0 +1,5 @@
+import Reflux from 'reflux';
+
+export default Reflux.createActions({
+  loadPolicies: {asyncResult: true}
+});
